Backend/requirements.txt
# Issue Tracker Backend — Python Dependencies
# Install with: pip install -r requirements.txt

# ─── Web Framework ──────────────────────────────────────────────────────────
fastapi==0.111.1
uvicorn[standard]==0.30.1

# ─── Database ───────────────────────────────────────────────────────────────
sqlalchemy==2.0.30
asyncpg==0.29.0
alembic==1.13.1

# ─── Validation ─────────────────────────────────────────────────────────────
pydantic==2.7.3
pydantic-settings==2.3.4

# ─── Environment ────────────────────────────────────────────────────────────
python-dotenv==1.0.1

# ─── Testing ────────────────────────────────────────────────────────────────
pytest==8.2.2
pytest-asyncio==0.23.7
httpx==0.27.0                  # async test client for FastAPI
